requirements.txt
pandas==2.1.4
numpy==1.26.3
pyodbc==5.0.1
python-dotenv==1.0.1
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.109.0